+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/directb2s/BackglassServiceClient.java
@@ -16,6 +16,8 @@ import org.springframework.http.ResponseEntity;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
+import java.util.Arrays;
+import java.util.List;
 
 /*********************************************************************************************************************
  * DirectB2S
@@ -38,6 +40,14 @@ public class BackglassServiceClient extends VPinStudioClientService {
 
   public DirectB2SData getDirectB2SData(int gameId) {
     return getRestClient().get(API + "directb2s/" + gameId, DirectB2SData.class);
+  }
+
+  public DirectB2SData getDirectB2SData(int emulatorId, String name) {
+    return getRestClient().get(API + "directb2s/" + emulatorId + "/" + name, DirectB2SData.class);
+  }
+
+  public List<DirectB2S> getBackglasses() {
+    return Arrays.asList(getRestClient().get(API + "directb2s", DirectB2S[].class));
   }
 
   public DirectB2ServerSettings getServerSettings(int emuId) {